Moved tests from old view model
